+++ apps/karman/karman_viewer.h
@@ -85,7 +85,7 @@ private:
     // Simulation variables
     size_t            m_steps;
     int               m_mnups;
-    int               m_num_particles;
+    size_t            m_num_particles;
     std::vector<Real> m_mean_velocity;
     int               m_forcing;
     Real              m_Re = 80.0; // Reynolds number
